with_info_option_without_monitor_returns_false
