+++ lumen_runtime/src/otp/erlang/demonitor_2/test/with_reference/with_flush_and_info_options.rs
@@ -3,22 +3,10 @@ mod with_monitor;
 use super::*;
 
 use crate::otp::erlang::monitor_2;
-use crate::process::SchedulerDependentAlloc;
 
 #[test]
 fn without_monitor_returns_false() {
-    with_process_arc(|monitoring_arc_process| {
-        let reference = monitoring_arc_process.next_reference().unwrap();
-
-        assert_eq!(
-            native(
-                &monitoring_arc_process,
-                reference,
-                options(&monitoring_arc_process)
-            ),
-            Ok(false.into())
-        )
-    });
+    with_info_option_without_monitor_returns_false(options);
 }
 
 fn options(process: &Process) -> Term {
